Fixed character movement, added, char writing, char waving
