--- Assets/BoxSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: adb1b9ebf7fa64253b5697fb6a3998f4
