--- Assets/GameManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d2581ec69af574c40883cb9dcfbf784e
