--- Assets/MoveAndDestroy.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c9c6e7d81bacb744da1616b9a97546a8
