+++ Assets/Scenes/VRScene.unity
@@ -18305,7 +18305,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!65 &1509134845
 BoxCollider:
   m_ObjectHideFlags: 0
