--- Assets/ScoreManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 51ff5099add214e07919995b1f65414f
