+++ Assets/Scripts/BoxSpawner2D.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c51716ff56da8433cb80c20a30eb37ec
