+++ Assets/Scripts/Load2dScene.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 25832e64f07b14e30948729d4aea0040
