+++ Assets/Scripts/ScoreManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 51ff5099add214e07919995b1f65414f
