+++ Assets/Scripts/WomanSpawner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2af2ccb2e830045218d90b1d9d009d4c
