--- Assets/WomanController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dfd80078a586e4ab9996191bc26b72c8
